Change float type variables to double.
We don't need float as we are not too worried about
space savings and we want to avoid rounding errors.

+++ src/main/java/seng201/team56/models/Cart.java
@@ -5,7 +5,7 @@ package seng201.team56.models;
  * @author Sean Reitsma
  */
 public class Cart {
-    private float speed;
+    private double speed;
     private int size;
     private ResourceType resourceType;
     private int resourceAmount;
@@ -17,7 +17,7 @@ public class Cart {
      * @param size the size of the cart
      * @param resourceType the ResourceType the cart can hold
      */
-    public Cart(float speed, int size, ResourceType resourceType) {
+    public Cart(double speed, int size, ResourceType resourceType) {
         this.speed = speed;
         this.size = size;
         this.resourceType = resourceType;
